Moved conversation endpoints into their own subclient
